+++ finance/src/sgpp/finance/application/HullWhiteSolver.cpp
@@ -58,7 +58,7 @@ HullWhiteSolver::~HullWhiteSolver() {
 
 void HullWhiteSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
-  this->levels = level;
+  this->levels = static_cast<size_t>(level);
 
   this->myGrid = new base::LinearBoundaryGrid(BoundingBox);
